Add optional support for VSG run-time visualization
